[ash] Allow readonly calendar access through OAuth with kSignin consent
Ash requests this scope at a time where the user is signed in but not
syncing. Since it is not whitelisted, it hits a CHECK and crashes.

It's not clear how it is possible for the user to be in this state, but
to prevent the crash, this CL adds the calendar scope to the list of
allowed scopes on ash.

Fixed: 1411735
Change-Id: I265b189fca3cf263feaf49f83326d89bbe773984
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4209215
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100310}