[Passwords] Improve the criterion of auto-submission for a combined form
Combined form is a <form> element (or the set of all <form>less inputs)
that encloses several user form (e.g. signin and signup on
https://www.uworld.com/app/index.html#/login/).
A submit button is good signal of the end of a user form. For example:
      <form>
        <div>Login</div>
          <input type='email' id='login_username'>
          <input type='password' id='login_password'>
          <input type='submit' id='login_submit'>
        <div>Create an account</div>
          <input type='email' id='signup_username'>
          <input type='text' id='signup_full_name'>
          <input type='password' id='signup_password'>
          <button type='submit' id='signup_submit'>
        <div>Forgot password?</div>
          <input type='email' id='recover_username'>
          <span>Request a recovery link</span>
      </form>

Currently, the heuristic to infer whether a form is ready for
auto-submission checks whether there are other input fields after
the focused input field. If no, then the form is ready for
auto-submission (FocusController::NextFocusableElementForIME).

After this CL, the heuristic doesn't look beyond a submit button (
<input type='submit'>, <button type='submit'>, <input type='image'>).

Bug: 1408040
Change-Id: I44205fd3d9045633e8490b20ae41ecb83431aa0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4205107
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100341}