Add test for corrections on focusable, aria-hidden nodes.
It is not allowed to set aria-hidden=true on nodes that can be focused
due to their role, or because they have a tabindex property. Authors may
still do that, anyway.

There are some cases where we are correcting bad authoring, by setting
a name to focusable nodes regardless of them being aria-hidden. We
consider that it's better that any elements that can be focused expose
a name.

These corrections are not always explicit in the code, some are side
effects of the conditions we used, the order in which they are applied,
etc. When we made corrections explicit, we followed a best effort
approach: we correct as long as we prevent using expensive operations.
For these reasons, we don't correct all the cases in the included test.

We correct these cases as a side effect. We check if a node has the
aria-hidden=true attribute and is focusable, but we don't check
aria-hidden on its ancestors:
* "outer aria-hidden": parent of the focusable node is aria-hidden.
* "outer aria-hidden, grandchildren": parent of the focusable node is
  aria-hidden, text node is a grandchild of the focusable node.

We correct this case explicitly because we check if the node has the
aria-hidden=true attribute and is focusable at the same time:
* "direct aria-hidden": focusable node is aria-hidden.

We correct this case as a side effect: the node containing text is not
focusable, but it's not directly aria-hidden either:
* "direct aria-hidden, grandchildren": focusable node is aria-hidden
  and text node is a grandchild of the focusable node.

We don't correct this case because we don't want to check if any of
the ancestors is focusable:
* "inner aria-hidden, grandchildren": child of focusable is not
  focusable and is aria-hidden.

With this test, we want to increase awareness of these corrections
and track whether further changes affect them.

Bug: 1330931
Change-Id: Ic02b9d24301daf3a7cb4800b71530dc65e0a3592
AX-relnotes: n/a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3804568
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Jacobo Aragunde Pérez <jaragunde@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1100369}