[Autofill] Add TestAutofillClientInjector
This CL adds a new RAII type TestAutofillClientInjector that replaces
a WebContent's AutofillClient. It's API follows that of
TestAutofillManagerInjector.

This class injector was originally intended to return nullptr in
AutofillClient::GetDownloadManager(). We're disabling the
AutofillDownloadManager via the fieldtrial testing config.

So for now, the TestAutofillClientInjector is not used, but it's
still likely useful for future tests browser tests.

This was meant as a preparatory CL for moving AutofillDownloadManager
from AutofillManager to AutofillClient: crrev.com/c/4165604.

Bug: 1394786, 892612
Change-Id: If77d032c7d8dbbfddc4283c4737af39c98165f60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4183770
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100376}