[WebAuthn] Add PasskeyCredential type to the Password Manager
Currently, WebAuthnCredentialsDelegate provides autofill::Suggestions
to the password autofill code to represent passkeys. This is not an
appropriate use of a type that is intended to be part of the autofill
UI.

This change moves the type currently used in Touch To Fill,
TouchToFillWebAuthnCredential, into components/password_manager and
renames it to PasskeyCredential. PasskeyCredential becomes used to
represent passkeys in password manager code until they need to be
converted to another type for display (Suggestion for autofill,
and the Java passkey class for Touch To Fill).

This also replaces the term 'WebAuthn credential' within the
password manager with 'passkey', since this is becoming the more
recognizable term, and the only WebAuthn credentials that can be
shown in autofill are passkeys.

Bug: 1355568
Change-Id: If12f2cc415806031a924cd85ad82fd90c9a746ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4195487
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100391}