Add GPUAdapter::SupportsMultiPlanarFormats()
"multi-planar-formats" feature is not in GPUFeatureName list
but currently GPUSupportedFeatures manages whether it is
supported.

"multi-planar-formats" is not needed to be exposed to JavaScript
so managing the compatibility in GPUSupportedFeatures may be weird.

This commit moves "multi-planar-formats" feature compatibility
management from GPUSupportedFeatures to GPUAdapter by adding
SupportsMultiPlanarFormats() to GPUAdapter.

Bug: 1379384
Change-Id: I394abdfb8ec5b264b6161f1650d96d254c479709
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4205056
Commit-Queue: Takahiro <hogehoge@gachapin.jp>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100467}