capture_mode: Use correct timestamps for video frames
When providing the encoder output to the muxer, we used to
use the timestamp value that the encoder provides. However,
that timestamp is a TimeDelta (which is a time difference
since the very first video frame in the beginning of the
capture).

The video frame already has a `reference_time` in its metadata,
which is meant for use as the time the frame was locally generated.
This CL uses this value instead of the previously-mentioned
timestamp.

Fixed: b/266786618
Test: Manually

Change-Id: I14c580aeda6647d1933e4182327c56e943f9638d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4201270
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Michele Fan <michelefan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100486}