Remove MessagePopupCollection hot mode feature
UX decided to remove this feature. There are a few bugs
associated with this feature that hastened this decision.

1. Large popups can dip below the shelf when their tops
are aligned with previously removed smaller popups.
2. Once a popup is outside of the work area bounds we can
enter a loop where the popup is removed because of being
out of bounds and then reshown because the notification still
exists.

Bug: b:267322636
Change-Id: Ia772b61839d6f7870f263dadf04774751e9e4c63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4211451
Commit-Queue: Ahmed Mehfooz <amehfooz@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100588}