BugFix: Make context current or mark lost for cached textures
When working with overlays, there are no guarantees that GL context
is current as they don't need GL. When a shared image representation
is destroyed, if it is the last representation of the associated
backing, the shared image backing also gets destroyed. In this case,
we have an overlay representation that was possibly the last
representation on ozone backing that got destroyed. This led to ozone
backing's destruction which in turn led to destoying its cached GL
textures (caching only done as a gpu driver bug workaround). Now,
destruction of a GL texture (glDeleteTexture) requires GL context
being current which is not the case. As a result, we now need to
check and either make context current or mark for context loss when
destroying overlay representation.

Ideally, we shouldn't need to cache the textures in backing and find
a better way to handle making context current.

Bug: 1408399
Change-Id: I9b2bf9686f8e923ad7caa5b683128e626edd9a84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4214873
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Kevin Haslett <khaslett@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100589}