[ZPS] Update ZeroSuggestCacheService tests & tweak `GetSuggestResults`.
This CL updates the unit tests for `ZeroSuggestCacheService`, while also
cleaning up the implementation of `CacheEntry::GetSuggestResults()` to
improve the readability of the code.

In addition to improving readability, this refactor also eliminates the
implicit assumption, which the pre-existing code depended on, that
`SearchSuggestionParser::ParseSuggestResults()` would NOT mutate the
input `results` argument on failure.

Bug: 1351224
Change-Id: I8c88a07fa91eceaec7610340d7eb8e13185fcfd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4215552
Commit-Queue: Khalid Peer <khalidpeer@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100644}