Check that H264NALU size is unsigned before creating spans
base::span will shortly require that the size field is of
type size_t and is thus unsigned.

h264decoder uses the `size` field of the H264NALU structure,
which is of type `off_t`, to create spans in multiple places.

To ensure that this is safe, and will continue to compile
once base::span checks have been enabled, a series of
base::checked_casts are added in this CL to ensure that
the current H264NALU size is really unsigned before
passing to base::span.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Change-Id: I8dab21b55bdcee8e34ae470ceccf8c7112b56809
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219876
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100647}