float: Move ScopedFloatDragHelper to new class
It turns out we will need this class for the duration of overview, see
bug for more details.

For that, we will need ScopedFloatDragHelper to be owned by the session
instead.

This CL is a pure refactor, so that the interesting changes in
following patches are easier to diff against. The class is still owned
by `OverviewWindowDragController`, but that will be addressed in the
following CL.

Test: existing tests
Bug: b/266947129
Change-Id: Ifae40bb92f3a990d0588750003046cd269a1d2ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219651
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100665}