Use size_t for kChunks constant in spdy test.
base::span will shortly require that size field is of
type size_t and is thus unsigned.

Previously the SpdyNetworkTransactionTest.PartialWrite
passed a signed value into base::make_span.

This CL changes this to a size_t value. Since this is a
test and it is known that chunks never exceeds 5, it is
also safe to implicit cast each chunk sequence number
into the int `sequence_number` field without risk of
precision loss.

This CL ensures that when span checks are enabled in a
future CL, the code will continue to compile.

No behavior change is intended by this CL.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Change-Id: I0c9a78d4550dec8fcd2235cd298a138664a2dc58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219088
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100676}