Fix assertions when CCT redirects set a target package
When a target package is sent for an outgoing intent from a CCT it
changes the resolveInfos, invalidating the cached result. However, in
this case we don't actually need the resolveInfos because we don't need
to check if we should avoid the disambiguation dialog as a package is
set on the Intent.

Bug: 1363279
Change-Id: I8c8c62c5f38b0ad55b16399021257782058a9619
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219034
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100698}