Update SignatureVerifierTest to use subspans
This CL removes some of the pointer arithmetic and replaces it
with calls to subspan to obtain the data needed for the test.

This also replaces a signed '1' with an unsigned '1u' as this
will be needed once base::span enforces unsigned values
for sizes in a future CL, and might as well be in this CL.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Change-Id: Id2cbc0f7c48a02cf802706f4231aa5d0b26f6b36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219321
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100736}