Pass size_t length into base::make_span in extensions.
Changes landing shortly enforce that base::span always takes an
unsigned value that fits into a size_t.

A check already exists that bytes_read is > 0 and this CL adds
a static_cast to size_t so it can be passed to base::make_span
without a compiler error in the future.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Change-Id: I71982355aeba96a889ad304c8ac16e1f47661bf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220471
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100743}