Attribution reporting API should be notified when fenced frame beacon is started
Due to recent change crrev.com/c/4144243, fenced frame beacons may be
cached and sent later when url mapping is ready. It's not guaranteed
that the initiator render frame host is still valid when the beacon is
sent. Therefore, Attribution Reporting API should be notified when the
beacon is started and the initiator render frame host is still valid.
This is also necessary so that we know which navigations to keep track
of as the actual beacon may be sent after the navigation finishes.

The NotifyFencedFrameReportingBeaconSent and
NotifyFencedFrameReportingBeaconData interfaces are removed from
AttributionHost and fenced frame reporting will call the interfaces in
AttributionDataHostManager which has longer lifetime.

The integration with fenced frame is in crrev.com/c/4208790.

Bug: 1408902
Change-Id: I0d42e41845371c72d76264f7322a00566449438d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4209475
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Commit-Queue: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100751}