Pass size_t size into base::make_span in mojo.
base::span will shortly require that length is unsigned and fits
into a size_t type.

The while loop already verifies that bytes_read > data_offset.

This CL can therefore safely static_cast this value to a size_t
and pass it to the base::make_span function.

This avoids a compile error in the future.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Change-Id: I8d519ac9db7346cd27aa7e0ccb03223da6dbcad2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219320
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1100766}