Overview Desk Navigation: Prevent overview grid bounds changing
It was discovered that when doing a short desk swipe and animating back
to the original desk, the overview grid items would all shift, resulting
in different positions as compared to when the user first enters
overview. This exposed a pre-existing bug in `overview_utils` where the
overview grid bounds could be calculated with respect to the hotseat,
even when the device is not in tablet mode.

This CL ensures that we add hotseat insets to the overview bounds only
only if we are in tablet mode, which prevents unnecessary shifts in
overview item positioning.

BUG=b:264958030
TEST=manual, updated existing test

Change-Id: I5c39f403ab7aed7d63eee089c98b9be029036486
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4210861
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Richard Chui <richui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100776}