Add Fake InputDevice instance for Touchscreen.
Due to the limitation of wayland protocol, there's no way to obtain
input device info from the compositor now.

We're making fake InputDevice instances for mouse/keyboard already,
so this extends it to touch devices for short term workaround.
This is just for the mitigation of the issue we see in Lacros.

For the long term, we probably need extended APIs and so further
discussions will be needed.

BUG=1405014
TEST=Ran on DUT.

Change-Id: Idb0c2de9c553f01bf38e92a8f29ce3fea91f7cea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4191616
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100827}