Overview Desk Navigation: Remove NoWindowsWidget animation
When entering overview with no items, we show a NoWindowsWidget. For
persisting overview state while switching desks, there is still a
default widget animation on the RoundedLabelWidget, causing the widget
to not be shown yet when taking the desk switch animation screenshot.

Similarly to https://crrev.com/c/4193198, we need to disable the default
widget animations for items to immediately show when for certain
overview enter/exit types. This CL introduces an extra parameter the
RoundedLabelWidget init parameters in order to do this, since this
widget animation property must be set after we initialized the widget,
but before we call `Show()`.

Also included some minor refactoring for when we don't want to animate entering overview.

BUG=b:264958030
TEST=manual, on test device

Change-Id: I0fc847a03c6776186564e0ad8bece288028cbd16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4212109
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Richard Chui <richui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100845}